+++ scripts/utils/config.js
@@ -8,6 +8,7 @@ import { AutoPopulateSettingDialog, CombatSettingDialog, GlobalSettingDialog, Ho
 // import ColorSetting from "/modules/colorsettings/colorSetting.js";
 import Picker from "/modules/colorsettings/lib/vanilla-picker.min.mjs";
 import API from "/modules/colorsettings/api.js";
+import { CPRActionsDialog } from "../components/dialog/CPRActionsDialog.js";
 
 export const BG3CONFIG = {
     // UI Constants
@@ -205,87 +206,6 @@ export const BG3CONFIG = {
         "--bg3-tooltip-text-secondary-color": "var(--bg3-text-secondary-color)",
         "--bg3-tooltip-component-color": "#aaaaaa",
         "--bg3-tooltip-border-size": "var(--bg3-border-size)"
-    },
-
-    COMBATACTIONDATA: {
-        "0-0": {
-            "uuid": null,
-            "name": "Grapple",
-            "img": "icons/magic/control/buff-strength-muscle-damage-red.webp",
-            "description": "<p>When you want to grab a creature or wrestle with it, you can use the Attack action to make a special melee attack, a grapple. If you're able to make multiple attacks with the Attack action, this attack replaces one of them. The target of your grapple must be no more than one size larger than you, and it must be within your reach.</p><p>Using at least one free hand, you try to seize the target by making a grapple check, a Strength (Athletics) check contested by the target's Strength (Athletics) or Dexterity (Acrobatics) check (the target chooses the ability to use). You succeed automatically if the target is incapacitated. If you succeed, you subject the target to the grappled condition (see the appendix). The condition specifies the things that end it, and you can release the target whenever you like (no action required).</p>",
-            "type": "feat",
-            "activation": {
-                "type": "action",
-                "cost": null,
-                "condition": ""
-            },
-            "sortData": {}
-        },
-        "1-0": {
-            "uuid": null,
-            "name": "Shove",
-            "img": "icons/skills/melee/unarmed-punch-fist-white.webp",
-            "description": "<p>When you want to grab a creature or wrestle with it, you can use the Attack action to make a special melee attack, a grapple. If you're able to make multiple attacks with the Attack action, this attack replaces one of them.</p><p>The target of your grapple must be no more than one size larger than you and must be within your reach. Using at least one free hand, you try to seize the target by making a grapple check instead of an attack roll: a Strength (Athletics) check contested by the target's Strength (Athletics) or Dexterity (Acrobatics) check (the target chooses the ability to use). You succeed automatically if the target is incapacitated. If you succeed, you subject the target to the grappled condition. The condition specifies the things that end it, and you can release the target whenever you like (no action required).</p><p>Escaping a Grapple. A grappled creature can use its action to escape. To do so, it must succeed on a Strength (Athletics) or Dexterity (Acrobatics) check contested by your Strength (Athletics) check.</p><p>Moving a Grappled Creature. When you move, you can drag or carry the grappled creature with you, but your speed is halved, unless the creature is two or more sizes smaller than you.</p><h4 id=\"ShovingaCreature\">Shoving a Creature</h4><p>Using the Attack action, you can make a special melee attack to shove a creature, either to knock it prone or push it away from you. If you're able to make multiple attacks with the Attack action, this attack replaces one of them.</p><p>The target must be no more than one size larger than you and must be within your reach. Instead of making an attack roll, you make a Strength (Athletics) check contested by the target's Strength (Athletics) or Dexterity (Acrobatics) check (the target chooses the ability to use). You succeed automatically if the target is incapacitated. If you succeed, you either knock the target prone or push it 5 feet away from you.</p>",
-            "type": "feat",
-            "activation": {
-                "type": "action",
-                "cost": null,
-                "condition": ""
-            },
-            "sortData": {}
-        },
-        "1-2": {
-            "uuid": null,
-            "name": "Disengage",
-            "img": "icons/skills/movement/feet-winged-boots-blue.webp",
-            "description": "<p>If you take the Disengage action, your movement doesn't provoke opportunity attacks for the rest of the turn.</p>",
-            "type": "feat",
-            "activation": {
-                "type": "action",
-                "cost": null,
-                "condition": ""
-            },
-            "sortData": {}
-        },
-        "0-2": {
-            "uuid": null,
-            "name": "Dodge",
-            "img": "icons/equipment/shield/buckler-wooden-boss-lightning.webp",
-            "description": "<p>When you take the Dodge action, you focus entirely on avoiding attacks. Until the start of your next turn, any attack roll made against you has disadvantage if you can see the attacker, and you make Dexterity saving throws with advantage. You lose this benefit if you are incapacitated or if your speed drops to 0.</p>",
-            "type": "feat",
-            "activation": {
-                "type": "action",
-                "cost": null,
-                "condition": ""
-            },
-            "sortData": {}
-        },
-        "1-1": {
-            "uuid": null,
-            "name": "Dash",
-            "img": "icons/skills/movement/figure-running-gray.webp",
-            "description": "<p>When you take the Dash action, you gain extra movement for the current turn. The increase equals your speed, after applying any modifiers. With a speed of 30 feet, for example, you can move up to 60 feet on your turn if you dash.</p><p>Any increase or decrease to your speed changes this additional movement by the same amount. If your speed of 30 feet is reduced to 15 feet, for instance, you can move up to 30 feet this turn if you dash.</p>",
-            "type": "feat",
-            "activation": {
-                "type": "action",
-                "cost": null,
-                "condition": ""
-            },
-            "sortData": {}
-        },
-        "0-1": {
-            "uuid": null,
-            "name": "Hide",
-            "img": "icons/containers/barrels/barrel-open-brown-red.webp",
-            "description": "<p>When you take the Hide action, you make a Dexterity (Stealth) check in an attempt to hide, following the rules for hiding. If you succeed, you gain certain benefits.</p><p>Combatants often try to escape their foes' notice by hiding, casting the invisibility spell, or lurking in darkness.</p><p>When you attack a target that you can't see, you have disadvantage on the attack roll. This is true whether you're guessing the target's location or you're targeting a creature you can hear but not see. If the target isn't in the location you targeted, you automatically miss, but the DM typically just says that the attack missed, not whether you guessed the target's location correctly.</p><p>When a creature can't see you, you have advantage on attack rolls against it. If you are hidden--both unseen and unheard--when you make an attack, you give away your location when the attack hits or misses.</p>",
-            "type": "feat",
-            "activation": {
-                "type": "action",
-                "cost": null,
-                "condition": ""
-            },
-            "sortData": {}
-        }
     }
 };
 
@@ -466,6 +386,7 @@ export function updateSettingsDisplay() {
     Hooks.on("renderSettingsConfig", (app, html, data) => {
         $('button[data-key="bg3-inspired-hotbar.menuExtraInfo"]').parents('div.form-group:first').remove();
         $('button[data-key="bg3-inspired-hotbar.containerAutoPopulateSettings"]').parents('div.form-group:first').remove();
+        $('button[data-key="bg3-inspired-hotbar.chooseCPRActions"]').parents('div.form-group:first').remove();
     });
 }
 
@@ -973,8 +894,7 @@ export function registerSettings() {
                 ui.BG3HOTBAR.components.portrait._renderInner();
             }
         },
-    });
-    
+    });    
 
     game.settings.register(BG3CONFIG.MODULE_NAME, 'showItemNames', {
         name: 'Show Item Names',
@@ -1068,6 +988,25 @@ export function registerSettings() {
         config: false,
         type: Boolean,
         default: true
+    });
+
+    game.settings.registerMenu(BG3CONFIG.MODULE_NAME, "chooseCPRActions", {
+        name: 'BG3.Settings.Menu.CPR.Name',
+        label: 'BG3.Settings.Menu.CPR.Label',
+        hint: 'BG3.Settings.Menu.CPR.Hint',
+        icon: "fas fa-cog",
+        type: CPRActionsDialog,
+        scope: 'world',
+        visible: () => {
+            return game.modules.get("chris-premades")?.active;
+        }
+    });
+
+    game.settings.register(BG3CONFIG.MODULE_NAME, 'choosenCPRActions', {
+        scope: 'client',
+        config: false,
+        type: Array,
+        default: ["9wbU6kYxfAaRFrbI", "ga6foNaesV3UJFKm", "eqOOv3smPuxTq7Xm", "pmn1iLabeps5aPtW", "nmkcJWUba7hyi5m5", "34jFXjMOseErle3M"]
     });
 
     game.settings.register(BG3CONFIG.MODULE_NAME, 'lockCombatContainer', {
@@ -1343,7 +1282,14 @@ export function registerHandlebars() {
 
     Handlebars.registerHelper('isdefined', function (value) {
         return value !== undefined;
-    });      
+    });
+    
+    Handlebars.registerHelper('ifIn', function(elem, list, options) {
+        if(list.indexOf(elem) > -1) {
+          return options.fn(this);
+        }
+        return options.inverse(this);
+    });
 
     /* Handlebars.registerHelper('check', function(fn, options) {
         // console.log(v1)